move pool to Apache folder
